TaskManager/requirement.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.0
django-widget-tweaks==1.5.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
